Tabe relationships

+++ src/tasks/tasks.service.ts
@@ -14,10 +14,12 @@ export class TasksService {
     private readonly taskRepository: Repository<Task>,
   ) {}
 
-  async getTasks(dto: TaskFilTerDto) {
+  async getTasks(dto: TaskFilTerDto, user: User) {
     const { status, search } = dto;
     const query: SelectQueryBuilder<Task> =
-      await this.taskRepository.createQueryBuilder('task');
+      this.taskRepository.createQueryBuilder('task');
+
+    query.where(`task.userId=:userId`, { userId: user.id });
 
     if (status) {
       query.andWhere('task.status=:status', { status });
@@ -34,15 +36,15 @@ export class TasksService {
     return tasks;
   }
 
-  async getTaskById(id: string): Promise<Task> {
-    const task = await this.taskRepository.findOneBy({ id });
+  async getTaskById(id: string, userId: string): Promise<Task> {
+    const task = await this.taskRepository.findOne({ where: { id, userId } });
     if (!task) {
       throw new NotFoundException('Task not found');
     }
     return task;
   }
 
-  async createTask(dto: CreateTaskDTO,user:User): Promise<Task> {
+  async createTask(dto: CreateTaskDTO, user: User): Promise<Task> {
     const { title, description } = dto;
     const task = new Task();
 
@@ -50,24 +52,27 @@ export class TasksService {
     task.title = title;
     task.description = description;
     task.status = 'OPEN';
-    task.user=user
+    task.user = user;
     await task.save();
- 
+
     delete task.user;
     return task;
   }
 
-  async editTask(id: string, status: string) {
-    const task = await this.taskRepository.update(id, {
-      status,
-    });
+  async editTask(id: string, status: string, userId: string) {
+    const task = await this.taskRepository.update(
+      { id, userId },
+      {
+        status,
+      },
+    );
     if (!task.affected) {
       throw new NotFoundException(`Task ${id} not found`);
     }
     return { updated: `task ${id} updated` };
   }
-  async deleteTask(id: string) {
-    const result = await this.taskRepository.delete(id);
+  async deleteTask(id: string, userId: string) {
+    const result = await this.taskRepository.delete({ id, userId });
     if (!result.affected) {
       throw new NotFoundException(`Task ${id} not found`);
     }
